+++ src/GameParser.hs
@@ -3,26 +3,60 @@ module GameParser
     ) where
 
 import Text.ParserCombinators.Parsec
+import qualified Data.Set as S
+import GameEngine
 
-parseGameFile :: String -> Either ParseError String
+parseGameFile :: String -> Either ParseError GameOptions
 parseGameFile str = parse gameFile "(unknown)" str
 
-gameFile :: CharParser () String
+gameFile :: CharParser () GameOptions
 gameFile = do 
     gameBeginToken
-    str <- manyTill anyChar (try gameEndToken)
-    return str
+    go <- headersSection
+    _ <- many anyChar
+    return go
+
+headersSection :: CharParser () GameOptions
+headersSection = do
+    spaces
+    string' "<Headers"
+    go <- S.fromList <$> (manyTill tryOptions (try (spaces *> string "/>")))
+    return go
+      where 
+        tryOptions = choice [
+            GameVersion <$> tryOption "Game Version",
+            GameName <$> tryOption "Game Name",
+            PlayerCapacity <$> tryOptionInt "Player Capacity",
+            StartingLocation <$> tryOption "Starting Location",
+            EndingLocation <$> tryOption "Ending Location"
+            ] 
+        tryOption name = try (readOption name)
+        tryOptionInt name = try (readOptionInt name)
+        readName name =  
+            do 
+                spaces
+                string' name
+                whiteSpaces
+                _ <- char '='
+                whiteSpaces
+        readOption name =
+            do 
+                readName name
+                str <- between (char '"') (char '"') (many (noneOf "\""))
+                whiteSpacesAndEol
+                return str
+        readOptionInt name = 
+            do 
+                readName name
+                str <- read <$> (many (alphaNum))
+                whiteSpacesAndEol
+                return str
 
 gameBeginToken :: CharParser () ()
 gameBeginToken = do
-    whiteSpacesAndNl 
-    _ <- string "Game Begin"
-    whiteSpaces
-    eol'
-    return ()
-
-gameEndToken :: CharParser () ()
-gameEndToken = string "Game End" >> return ()
+    spaces 
+    _ <- string "<Game"
+    whiteSpacesAndEol
 
 eol :: CharParser () String
 eol = try (string "\n\r")
@@ -34,8 +68,10 @@ eol = try (string "\n\r")
 eol' :: CharParser () ()
 eol' = eol >> return ()
 
-whiteSpacesAndNl :: CharParser () ()
-whiteSpacesAndNl = many (oneOf " \t\r\n") >> return ()
+string' :: String -> CharParser () ()
+string' str = string str *> return ()
 
 whiteSpaces :: CharParser () ()
-whiteSpaces = many (oneOf " \t")  >> return ()
+whiteSpaces = many (oneOf " \t\f\v")  *> return ()
+whiteSpacesAndEol :: CharParser () ()
+whiteSpacesAndEol = whiteSpaces *> eol'