+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog for Story-Teller-Engine
 
+## 0.2.0.0
+- Change begin and end token to < />
+- Game Options parseable!
 ## 0.1.1.0
 - Selecting game file
     - no exceptions handling